+++ CHANGELOG.md
@@ -18,6 +18,7 @@ npm install --global skyliner # For using the CLI globally
 
 ## v0.2 (unreleased)
  - Yield `null` if no language definition could be found
+ - Make compatible with [Rollup.js](https://https://rollupjs.org/guide/en/) and other bundlers
 
 
 ## v0.1 (20th March 2021)